fix(chat): address non-blocking review findings from sidebar stacked rows
- Keep time/status visible on stacked rows for touch devices that are
  not phones (hover: none tablets): the archive button takes a static
  slot beside the time instead of overlaying it.
- Finish the New task rename: project rows in the new-chat dialog and
  the project-context breadcrumb now say 'New task'.
- Extract ProjectRowIcon and the two task status badge clusters into
  ProjectChatItemBadges.tsx (row file 545 -> 441 lines).
